// sanity/sanity.types.ts
export type SanityImageHotspot = {
  _type: 'sanity.imageHotspot'
  x?: number
  y?: number
  height?: number
  width?: number
}

export type SanityImageCrop = {
  _type: 'sanity.imageCrop'
  top?: number
  bottom?: number
  left?: number
  right?: number
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol
// ...
export type USER_JOURNAL_ENTRIES_QUERYResult = Array<{
  _id: string
  title: string | null
  content: Array<
    | {
        children?: Array<{
          marks?: Array<string>
          text?: string
          _type: 'span'
          _key: string
        }>
        style?: 'blockquote' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'normal'
        listItem?: 'bullet' | 'number'
        markDefs?: Array<{
          href?: string
          _type: 'link'
          _key: string
        }>
        level?: number
        _type: 'block'
        _key: string
      }
    | {
        asset?: {
          _ref: string
          _type: 'reference'
          _weak?: boolean
          [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
        }
        media?: unknown
        hotspot?: SanityImageHotspot
        crop?: SanityImageCrop
        alt?: string
        caption?: string
        _type: 'image'
        _key: string
      }
  > | null
  mood: 'happy' | 'neutral' | 'sad' | 'very-happy' | 'very-sad' | null
  createdAt: string | null
  aiGeneratedCategory: {
    title: string | null
    color: string | null
  } | null
}>
// Variable: JOURNAL_ENTRY_BY_ID_QUERY
// Query: *[  _type == "journalEntry"   && _id == $entryId][0]{  _id,  title,  content,  mood,  createdAt,  userId,  aiGeneratedCategory->{    title,    color  }}
export type JOURNAL_ENTRY_BY_ID_QUERYResult = {
  _id: string
  title: string | null
  content: Array<
    | {
        children?: Array<{
          marks?: Array<string>
          text?: string
          _type: 'span'
          _key: string
        }>
        style?: 'blockquote' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'normal'
        listItem?: 'bullet' | 'number'
        markDefs?: Array<{
          href?: string
          _type: 'link'
          _key: string
        }>
        level?: number
        _type: 'block'
        _key: string
      }
    | {
        asset?: {
          _ref: string
          _type: 'reference'
          _weak?: boolean
          [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
        }
        media?: unknown
        hotspot?: SanityImageHotspot
        crop?: SanityImageCrop
        alt?: string
        caption?: string
        _type: 'image'
        _key: string
      }
  > | null
  mood: 'happy' | 'neutral' | 'sad' | 'very-happy' | 'very-sad' | null
  createdAt: string | null
  userId: string | null
  aiGeneratedCategory: {
    title: string | null
    color: string | null
  } | null
} | null
// Variable: USER_JOURNAL_ENTRIES_WITH_DATE_RANGE_QUERY
// Query: *[  _type == "journalEntry"   && userId == $userId  && createdAt >= $startDate  && createdAt <= $endDate] | order(createdAt desc) {  _id,  title,  content,  mood,  createdAt,  aiGeneratedCategory->{    title,    color  }}
export type USER_JOURNAL_ENTRIES_WITH_DATE_RANGE_QUERYResult = Array<{
  _id: string
  title: string | null
  content: Array<
    | {
        children?: Array<{
          marks?: Array<string>
          text?: string
          _type: 'span'
          _key: string
        }>
        style?: 'blockquote' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'normal'
        listItem?: 'bullet' | 'number'
        markDefs?: Array<{
          href?: string
          _type: 'link'
          _key: string
        }>
        level?: number
        _type: 'block'
        _key: string
      }
    | {
        asset?: {
          _ref: string
          _type: 'reference'
          _weak?: boolean
          [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
        }
        media?: unknown
        hotspot?: SanityImageHotspot
        crop?: SanityImageCrop
        alt?: string
        caption?: string
        _type: 'image'
        _key: string
      }
  > | null
  mood: 'happy' | 'neutral' | 'sad' | 'very-happy' | 'very-sad' | null
  createdAt: string | null
  aiGeneratedCategory: {
    title: string | null
    color: string | null
  } | null
}>